chore: feedback improvements useMemo and types
